Adding a heap and vmhandle
The VMHandle serves the purpose of giving access to the VM to workers, without
explicitly handing the VM off, resulting in everything in the VM needing to be thread safe due
to the fiber being submitted across all workers.

The Heap exists to represent data that must be handed off between workers, like loaded modules
or function declarations.