Modifying hash and string implementation, working gcj
The hash implementation in warpspeed now creates a string pool
which uses the hash of the string as the key. This ensures that the
same constant string will result in the same key in the hash.

Previously each string had a unique index in the array, which
resulted in duplicate strings being stored with different indices.

With this change, the google code jam 2008 qualifier works as expected.