Switch to deref vs boxing, adding length.
Many calls to obtain data structures to work with
were unboxing the raw pointer, which returned ownership
to the functions and causing the functions to drop the
data structure (in this case hashmaps).

Switching to dereferencing keeps the function from dropping
the object, and better explains the intent (borrow the values
vs consume them).