Migrated handling of indentation to parser
The conversion of indentation to a parenthesized syntax used to occur in the
preprocessor. The preprocessor was not intelligent enough to handle newlines in data
structure declarations, making it very difficult to express structures such as maps in a
readable way.

Migrating this logic into the parser enables indentation-based logic to be able to handle these rules.