Creating DispError to accompany GenericError
GenericError is simply an alias to a Box<Error>, which
is a great catch-all error that allows grouping of all errors
returned back by various libraries.

DispError was created as a generic struct that implements
error, ensuring it can be used in this case as well. More fine-grained
error checking will probably be needed in the future, but this is
sufficient for now.

Elaborating on a config-value disp case (config_values.ds)

Continuing to add support for maps in the parser.