maps honor string keys correctly, gcj_2008 works
Maps had the problem again of using pointer values rather than
actual values to specify equality, resulting in maps containing
duplicate entries. This was fixed by reverting HashMaps to use the
string types directly. This incurs a performance penalty due to the
conversion back to Strings, but can be addressed at a later time.

this fixes the final issue with the gcj_2008_2 code, and it returns
the correct answer.