Enabling a standard library
Have the ability to load multiple files in the compilation of a program,
using that logic to load a lib.ds file before all else, allowing for the kernel of
a standard library.

Including a refactor of the scope / compiler relationship, to enable
declaring functions on a global level that can be consumed by modules.