bugfix: functions compiling when called in nested scope
The scope object previously was the object traversed to find all functions
which should be compiled. This resulted in functions that were only called
in bodies of functions to not be detected by the llvm builder.

The function declarations are now stored in the top-level compiler object,
and the builder is now using that map to find all declarations.