Small refactor of CompilerData
moving the compiler workflow to workflow.rs.
moving CompilerData to compiler.rs.

The motivation for these changes it to ensure small units of code with
specific purposes. CompilerData will eventually be the global compiler
context, and should be moved out of codegen since it plays a much
larger part (e.g. being a store for macros)

Making space for that by moving the workflow of the compiler into
workflow.rs.