Refactored llvm_codegen, introduced LLVMTypeCache
Continuing an ongoing refactor to modularize the files a bit more. The
general idea is to lift concepts that are agnostic of any particular step
(such as types) into top level modules. For example, the types module
is being up with this philosophy in mind.

The other design change is to better abstract between LLVM and disp's
intentions. This will better enable Disp to harness other backends if it is
desired, and to better document what concepts need to be implemented
by a backend.

In addition, experimenting a bit with merging llvm_codegen and
llvm_builder. The motivation for separating the two was to enable
a more flexible builder interface (it started as a way to enable setting
types after a function declaration). However, this has been satisfied by the
type check phase of the compier, and the code could be simplified by merging
the two modules together.

In addition to the code complexity, abstractions are leaking into
llvm_codegen, such as the LLVM type references.

The work in this area is started by declaring the builder / context objects
in llvm_codegen, and passing them into builder.