Fixed the bytes struct passed in as a pointer, fixes get_bytes bad index
There was an LLVM error around a struct value being used
as the object type when it should have been a struct pointer.

Added a hack for now to return back the array as a pointer
(which is appropriate, as an array could be allocated on
the heap or on the stack and using a pointer enables
both).

This looks to have fixed a bug with an incorrect i64 being passed
into the get_bytes method.