Bytes are working
Fixed a bug where the length of the array struct was incorrect: This was due
to defining the first element in the array struct as a
rust array, which also includes a length usize value.

This resulted in the bug being hard to find: despite the fact that LLVM
initializing a c-style array rather than a Rust array, the
additional length value that was set during the creation of
the struct resulted in the result looking to LLDB and the language as
a Rust array.

Bytes works as expected in all scenarios now.