Added Polymorphic Unification to inference
Unification with Polymorphic values was not possible in the existing
implementation, due to consuming concrete types and not being able to
express generic values.

In order to do so, a wrapping enum was introduced which allows one
to declare a Generic, a Literal, and an "Any" value. This is a definition
that is able to consume type variables as the qualifiers of generics.

Constraints on generic type variables can be declared as with any
other type variable constraint: retrieving type variables and adding
constraints specific to them.

By creating the wrapper type, this also eliminates the need for the custom
unification trait of types: any generic declaration can now be used in
type inference.