Have array declarations working.
Array declarations are now properly identified as the correct type,
as well as can be used in generic functions such as (get).

LLVM arrays, when indexed via GEP, must also include a reference to the
zero index beforehand, before passing the correct index.

There is an outstanding LLVM error around pushing an array type into
a point of the underlying data type. It seems to be working, but there
could be ramifications.