started to add hash function, cleaned up macros
currently there is a pattern where authoring a macro with the
intuitive syntax does not result in a clean, expected representation.

macros directly substitute the macro invocation expression
with the token in question. as a result, authoring a macro like this:

macro! foo [var]
    print var

results in the a replacement macro invocation of a block
where it should probably act as a single expression:

macro! foo [var] <(print var)>

foo "x"

becomes

<<(foo "x")>>

(1st block the main function itself, inner block is coming from the
macro).

for now, removing the pattern of using blocks for each macro and instead
making it an inline invocation. In the future we probably need to do
something like unpack macro bodies.